pyproject.toml
[tool.poetry]
name = "workshop-telegram-bot"
version = "0.1.0"
description = ""
authors = ["keltio"]
license = "of course"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.11.0"
pillow = "^10.4.0"
python-dotenv = "^1.0.1"
aiosqlite = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
